requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Flask-Login==0.6.3
Pillow==10.2.0
gunicorn==21.2.0
